### 2024-09-03 #653 @maelle
emodnet.wfs: Access EMODnet Web Feature Service data through R 
Submitting Author Name: Maëlle Salmon
Submitting Author Github Handle: @maelle
Other Package Authors Github handles: (comma separated, delete if none) @salvafern, @annakrystalli
Repository:  https://github.com/EMODnet/emodnet.wfs/
Version submitted: 2.0.2
Submission type: Standard

Version accepted: TBD

---

```
Package: emodnet.wfs
Title: Access EMODnet Web Feature Service data through R
Version: 2.0.2
    person("Anna", "Krystalli", , "annakrystalli@googlemail.com", role = "aut",
    person("Salvador", "Fernández-Bejarano", , "salvador.fernandez@vliz.be", role = c("aut", "cre"),
    person("Thomas J", "Webb", , "t.j.webb@sheffield.ac.uk", role = "ctb"),
    person("European Marine Observation Data Network (EMODnet) Biology project", "European Commission's Directorate - General for Maritime Affairs and Fisheries (DG MARE)", , "bio@emodnet.eu", role = "cph"),
    person("VLIZ (VLAAMS INSTITUUT VOOR DE ZEE)", , , "info@vliz.be", role = "fnd"),
    person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = "aut",
  )
Description: Access and interrogate EMODnet (European Marine Observation and Data Network)
    Web Feature Service data through R.
    https://github.com/EMODnet/emodnet.wfs
    checkmate,
    cli,
    lifecycle,
    memoise,
    ows4R (>= 0.4),
    sf,
    whoami
    covr,
    httptest,
    mapview,
    skimr,
    testthis,
    withr
Config/Needs/readme: rerddap
    knitr
SystemRequirements: C++11, GDAL (>= 2.0.1), GEOS (>= 3.4.0), PROJ (>=
    4.8.0)
    eblondel/ows4R

```

The package allow interrogation of and access to EMODnet’s, European Marine Observation and Data Network, geographic vector data.

The target audience of the package are EMODnet users that might need programmatic access to EMODnet's geographic vector data.
The package allows to include EMODnet vector data into scientific pipelines without needing to manually explore and download data from the [EMODnet Geographic Viewer](https://emodnet.ec.europa.eu/geoviewer/).
The data covers seven disciplinary themes (bathymetry, geology, physics, chemistry, biology, seabed habitats and human activities).

No, to our knowledge emodnet.wfs is the only package that provides access to EMODnet data in R though the EMODnet Web Feature
Services.
The emodnet.wfs package was developed in collaboration with other EMODnet members.

There are in total three ways to access EMODnet data that complement each other and which we documented in [emodnet.wfs README](https://github.com/EMODnet/emodnet.wfs?tab=readme-ov-file#other-web-services):

- Some EMODnet data are also published in an [ERDDAP server](https://erddap.emodnet.eu). One can access these data in R using the rOpenSci [rerddap R package](https://docs.ropensci.org/rerddap/).

- This package emodnet.wfs uses [Web Feature Services](https://www.ogc.org/standard/wfs/), hence it is limited to getting vector data. EMODnet also hosts raster data that can be accessed via [Web Coverage Services (WCS)](https://www.ogc.org/standard/wcs/). The R package [EMODnetWCS](https://github.com/EMODnet/EMODnetWCS) makes these data available in R. We intend to also submit the latter to software review at one point (and we intend to rename it).

N/A

-   If you made a pre-submission inquiry, please paste the link to the corresponding issue, forum post, or other discussion, or tag the editor you contacted.
