### 2024-09-04 #653 @adamhsparks
Hi @maelle, I've got a NOTE when I run checks on your package with the documentation.

```r
─  checking Rd files ... [0s/0s] NOTE
   checkRd: (-3) emodnet_get_layers.Rd:40-41: Lost braces
       40 | \item{character string or character vector of length 1.
          |      ^
   checkRd: (-3) emodnet_get_layers.Rd:42-44: Lost braces
       42 | \item{character vector of length equal to the length of layers.
          |      ^
   checkRd: (-3) emodnet_get_layers.Rd:45-48: Lost braces
       45 | \item{named character vector. Each filter will be applied to the layer
```

Also, the tests seem to take quite a while to run. Locally I edited the DESCRIPTION file to include `Config/testthat/parallel: true` so it was a *bit* faster.

Could you check on the NOTE and see if you can fix it and I'll assign an editor in the meantime?
