### 2024-09-04 ddd531c
docs: fix braces
Thanks @adamhsparks
