### 2024-09-04 beaab9b
docs: fix braces
Thanks @adamhsparks
