### 2024-09-04 ab440e9
chore: normalize DESCRIPTION
