+++ DESCRIPTION
@@ -12,8 +12,8 @@ Authors@R: c(
     person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = "aut",
            comment = c(ORCID = "0000-0002-2815-0399"))
   )
-Description: Access and interrogate EMODnet (European Marine Observation and Data Network) 
-    Web Feature Service data through R.
+Description: Access and interrogate EMODnet (European Marine Observation
+    and Data Network) Web Feature Service data through R.
 License: MIT + file LICENSE
 URL: https://emodnet.github.io/emodnet.wfs/,
     https://github.com/EMODnet/emodnet.wfs
@@ -45,9 +45,11 @@ Suggests:
     testthat (>= 3.1.2),
     testthis,
     withr
-Config/Needs/readme: rerddap
 VignetteBuilder: 
     knitr
+Remotes: 
+    eblondel/ows4R
+Config/Needs/readme: rerddap
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
@@ -55,5 +57,3 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 SystemRequirements: C++11, GDAL (>= 2.0.1), GEOS (>= 3.4.0), PROJ (>=
     4.8.0)
-Remotes: 
-    eblondel/ows4R
