### 2024-09-04 d590025
test: run tests in parallel
thanks @adamhsparks
