+++ DESCRIPTION
@@ -51,6 +51,7 @@ Remotes:
     eblondel/ows4R
 Config/Needs/readme: rerddap
 Config/testthat/edition: 3
+Config/testthat/parallel: true
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
