### 2024-09-04 12720b2
chore: namespace globalVariables()
