### 2024-09-04 #653 @maelle
Thanks @adamhsparks!

What exactly do you mean by running checks on the package with the documentation? I changed the markup for those items but would like to re-run the checks as you did to ensure I did fix the problem.

I have added that field to DESCRIPTION, so the tests run faster now, thank you!
