### 2024-09-04 #653 @adamhsparks
@maelle, it was just `devtools::check()`. The most recent commits appear to have fixed the NOTE with the "{" in checks.
