### 2024-09-07 #653 @MargaretSiple-NOAA
- [X ] Is the reference index page clear (grouped by topic if necessary)?
---

Great documentation and package setup. Thank you for being thorough!

A couple very minor notes:
- I suggest changing the name of the intro vignette "emodnet.wfs Case Study: Accessing and mapping EMODnet data" to just "Accessing and mapping EMODnet data" and moving to top of article list.
- I'm not totally sure what tokens / playgrounds would look like for this package but I found good contribution guidelines on the repo and IMO those are more than sufficient for contribution information.
- It looks like there are just a couple bugs in the Issues but they are actively being worked on.
---
