### 2024-09-09 38ac5bd
docs: change title of main vignette
thx @MargaretSiple-NOAA
