+++ vignettes/emodnet.wfs.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "emodnet.wfs Case Study: Accessing and mapping EMODnet data"
+title: "Accessing and mapping EMODnet data"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{API details}
@@ -94,33 +94,26 @@ The column `service_name` shows services available, while  `service_url` has the
 
 ``` r
 seabed_wfs_client <- emodnet_init_wfs_client(service = "seabed_habitats_general_datasets_and_products")
-#> Loading ISO 19139 XML schemas...
-#> Loading ISO 19115 codelists...
 #> ✔ WFS client created successfully
 #> ℹ Service: "https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open/wfs"
 #> ℹ Version: "2.0.0"
-```
-
-``` r
 
 emodnet_get_wfs_info(wfs = seabed_wfs_client)
-#> # A tibble: 71 × 9
+#> # A tibble: 72 × 9
 #> # Rowwise: 
-#>    data_source service_name        service_url layer_name title
-#>    <chr>       <chr>               <chr>       <chr>      <chr>
-#>  1 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  2 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  3 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  4 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  5 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  6 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  7 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  8 emodnet_wfs seabed_habitats_ge… https://ow… art17_hab… 2013…
-#>  9 emodnet_wfs seabed_habitats_ge… https://ow… carib_eus… 2023…
-#> 10 emodnet_wfs seabed_habitats_ge… https://ow… biogenic_… Biog…
-#> # ℹ 61 more rows
-#> # ℹ 4 more variables: abstract <chr>, class <chr>,
-#> #   format <chr>, layer_namespace <chr>
+#>    data_source service_name             service_url layer_name title abstract class format layer_namespace
+#>    <chr>       <chr>                    <chr>       <chr>      <chr> <chr>    <chr> <chr>  <chr>          
+#>  1 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
+#>  2 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
+#>  3 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
+#>  4 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
+#>  5 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
+#>  6 emodnet_wfs seabed_habitats_general… https://ow… art17_hab… 2013… "Gridde… WFSF… sf     emodnet_open   
[…]
