### 2024-09-17 d98df33
docs: add rOpenSci software-review badge
