### 2024-09-27 #653 @LizHareDogs
Hi! I'm a newbie reviewer and this package looks great! I've worked my way through the examples in README.md successfully.

There is a file in the vignettes folder `emodnet.wfs.Rmd`, but I do not find it rendered. The index page for the package does not list any vignettes. There is not a docs file where the rendered vignette/user guide might live?

It looks like there is good info in the .Rmd file that I don't want to miss.

Please let me know if you need any more information!
