### 2024-09-28 #653 @adamhsparks
@LizHareDogs, fair question. That's this vignette, https://emodnet.github.io/emodnet.wfs/articles/emodnet.wfs.html. You can see the title here:https://github.com/EMODnet/emodnet.wfs/blob/d98df3354e7abb1528c3eb6cc5bdd64e6bc8b90b/vignettes/emodnet.wfs.Rmd#L2 and in the HTML version the source is indicated on the second line right under the title, "Accessing and mapping EMODnet data".

You can find the vignettes on the webpage under the "Articles" menu item.
