### 2024-09-28 #653 @MargaretSiple-NOAA
@LizHareDogs Ah yes, vignettes can be a little confusing. Unlike some of the other package materials, a knitted/rendered form of that Rmd does not live with the package when you've just cloned it. There's some more information about that (and useful info about vignettes and development) [here ](https://r-pkgs.org/vignettes.html#sec-vignettes-how-built).

As Adam mentioned, for most packages, the vignettes will be the same as the articles on the `pkgdown` page, but they _can_ be different because vignettes ship with a package and articles only live on the package's website. It looks like the articles are the same as the vignettes for emodnet.wfs, but if you did want to compare them, you should be able to knit the Rmd files yourself in the cloned repo.
