### 2024-09-28 #653 @LizHareDogs
Many thanks @adamhsparks and @MargaretSiple-NOAA for the explanations and the resources!
