### 2024-09-30 #653 @maelle
Thanks all! To make things worse, this vignette is pre-rendered following [Jeroen's blog post "How to precompute package vignettes or pkgdown articles"](https://ropensci.org/blog/2019/12/08/precompute-vignettes/)

- The source is [`vignettes/emodnet.wfs.Rmd.orig`](https://github.com/EMODnet/emodnet.wfs/blob/main/vignettes/emodnet.wfs.Rmd.orig)
- It is rendered using `knitr::knit("vignettes/emodnet.wfs.Rmd.orig", output = "vignettes/emodnet.wfs.Rmd")`.

We have a [reminder to update the vignette](https://github.com/EMODnet/emodnet.wfs/blob/d98df3354e7abb1528c3eb6cc5bdd64e6bc8b90b/R/emodnet.wfs-package.R#L45) at least when we release with [`devtools::release()`](https://devtools.r-lib.org/reference/release.html#details). Until now we haven't used that function but we have remembered to update it.
