### 2024-10-14 #653 @LizHareDogs
Thanks for understanding @MargaretSiple-NOAA! We had a great meeting about my draft review and our plans. We think we can finish by October 30.
Is there a way to give mentors in the champions program credit for participating in the review? @fmichonneau made  good points that I'd like to add with credit to him.
