### 2024-10-17 3be8e11
docs: remove duplicate test coverage badge
Thanks @robitalec
