### 2024-10-17 a32454d
docs: fix intro vignette position in navbar
