### 2024-10-17 c2eef34
docs: fix webshot dependency problem
