### 2024-10-17 f5b657e
docs: avoid duplicates in navbar
