### 2024-10-17 03b019c
docs: avoid printing URLs that are invalid on their own
