### 2024-10-17 2c2e9df
docs: add CITATION.cff
