+++ CITATION.cff
@@ -1,0 +1,478 @@
+# --------------------------------------------
+# CITATION file created with {cffr} R package
+# See also: https://docs.ropensci.org/cffr/
+# --------------------------------------------
+ 
+cff-version: 1.2.0
+message: 'To cite package "emodnet.wfs" in publications use:'
+type: software
+license: MIT
+title: 'emodnet.wfs: Access EMODnet Web Feature Service data through R'
+version: 2.0.2.9000
+doi: 10.14284/679
+abstract: Access and interrogate EMODnet (European Marine Observation and Data Network)
+  Web Feature Service data through R.
+authors:
+- family-names: Krystalli
+  given-names: Anna
+  email: annakrystalli@googlemail.com
+  orcid: https://orcid.org/0000-0002-2378-4915
+- family-names: Fernández-Bejarano
+  given-names: Salvador
+  email: salvador.fernandez@vliz.be
+  orcid: https://orcid.org/0000-0003-0535-7677
+- family-names: Salmon
+  given-names: Maëlle
+  email: msmaellesalmon@gmail.com
+  orcid: https://orcid.org/0000-0002-2815-0399
+preferred-citation:
+  type: manual
+  title: 'emodnet.wfs: Access EMODnet Web Feature Service data through R'
+  authors:
+  - family-names: Krystalli
+    given-names: Anna
+    email: annakrystalli@googlemail.com
+    orcid: https://orcid.org/0000-0002-2378-4915
+  - family-names: Fernández-Bejarano
+    given-names: Salvador
+    email: salvador.fernandez@vliz.be
+    orcid: https://orcid.org/0000-0003-0535-7677
+  - family-names: Salmon
+    given-names: Maëlle
+    email: msmaellesalmon@gmail.com
+    orcid: https://orcid.org/0000-0002-2815-0399
+  notes: R package version 2.0.2.9000. Integrated data products created under the
+    European Marine Observation Data Network (EMODnet) Biology project (EASME/EMFF/2017/1.3.1.2/02/SI2.789013),
+    funded by the by the European Union under Regulation (EU) No 508/2014 of the European
+    Parliament and of the Council of 15 May 2014 on the European Maritime and Fisheries
+    Fund
+  url: https://github.com/EMODnet/emodnet.wfs
+  doi: 10.14284/679
+repository-code: https://github.com/EMODnet/emodnet.wfs
+url: https://emodnet.github.io/emodnet.wfs/
+contact:
+- family-names: Fernández-Bejarano
+  given-names: Salvador
+  email: salvador.fernandez@vliz.be
+  orcid: https://orcid.org/0000-0003-0535-7677
+keywords:
+- biology
+- dataproducts
+- emodnet
+- marine-data
+- rstats
+- wfs
+references:
+- type: software
+  title: 'R: A Language and Environment for Statistical Computing'
+  notes: Depends
+  url: https://www.R-project.org/
+  authors:
+  - name: R Core Team
+  institution:
+    name: R Foundation for Statistical Computing
+    address: Vienna, Austria
+  year: '2024'
+  version: '>= 3.6.0'
+- type: software
+  title: checkmate
+  abstract: 'checkmate: Fast and Versatile Argument Checks'
+  notes: Imports
+  url: https://mllg.github.io/checkmate/
+  repository: https://CRAN.R-project.org/package=checkmate
+  authors:
+  - family-names: Lang
+    given-names: Michel
+    email: michellang@gmail.com
+    orcid: https://orcid.org/0000-0001-9754-0393
+  year: '2024'
+  doi: 10.32614/CRAN.package.checkmate
[…]
