### 2024-10-17 fcd842b
chore: make skimr a dependency for the website only
