+++ CITATION.cff
@@ -397,34 +397,6 @@ references:
   year: '2024'
   doi: 10.32614/CRAN.package.rmarkdown
 - type: software
-  title: skimr
-  abstract: 'skimr: Compact and Flexible Summaries of Data'
-  notes: Suggests
-  url: https://docs.ropensci.org/skimr/
-  repository: https://CRAN.R-project.org/package=skimr
-  authors:
-  - family-names: Waring
-    given-names: Elin
-    email: elin.waring@gmail.com
-  - family-names: Quinn
-    given-names: Michael
-    email: msquinn@google.com
-  - family-names: McNamara
-    given-names: Amelia
-    email: amcnamara@smith.edu
-  - family-names: Rubia
-    given-names: Eduardo
-    name-particle: Arino de la
-    email: earino@gmail.com
-  - family-names: Zhu
-    given-names: Hao
-    email: haozhu233@gmail.com
-  - family-names: Ellis
-    given-names: Shannon
-    email: sellis18@jhmi.edu
-  year: '2024'
-  doi: 10.32614/CRAN.package.skimr
-- type: software
   title: testthat
   abstract: 'testthat: Unit Testing for R'
   notes: Suggests
