### 2024-10-29 #653 @LizHareDogs
## Package Review

- [X ] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [X ] **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
- [X ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [X ] **Vignette(s):** demonstrating major functionality that runs successfully locally
- [*** ] **Function Documentation:** for all exported functions
- [X ] **Examples:** (that run successfully locally) for all exported functions
- [- ] **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
Partially: the community guidelines and contribution guidelines are not available.

#### Functionality

- [X ] **Installation:** Installation succeeds as documented.
- [X ] **Functionality:** Any functional claims of the software have been confirmed.
- [NA ] **Performance:** Any performance claims of the software have been confirmed.
- [X ] **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
- [X ] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 25

- [X ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

---

### Review Comments

I am reviewing as part of the Champions program. My mentor is François Michonneau (@fmichonneau) and I have credited him with some of the suggestions below.

Line numbers mentioned below are in the unrendered files.

#### README.md

##### Readthrough

From the perspective of someone who has not worked with this kind of data,  it would be helpful if the introduction (which starts at line 21) could provide a little more context. What is geographic
vector data? I'm new to this kind of data so it would be helpful to know in a less technical way what it is.

Line 26
"though" should be "through"

line 78
Not sure if I can use "View"

line 184
What does CRS stand for (it's mentioned a few times later)?

Line 313
what is an sf object? What is the difference between this and how the data were retrieved before?

##### Following instructions on Mac OS using Emacs, Emacs SPeaks Statistics

Line 78
I am not able to evaluate the View command on my system. This is not a problem with the package.

Otherwise, sample commands work as expected.

#### Vignette

All the commands in the vignette work for me.

Line 24 has the heading "data product." The meaning of this isn't clear to me. Will this be instructions on how to access a data product or how to create a data product?

Line 109 has this command:
map <- mapview(habitats_directive_layers, zcol = "habitat_description", burst = TRUE)

I am able to run this command, which opens a browser. I can't check the map with the screen reader, but the text on the screen seems appropriate.

In the rendered vignette at
the following error is displayed for the mapview command:

```
#> Error in loadNamespace(name): there is no package called 'webshot'
```

#### devtools

No problems with:
* `check()`
* `test()`
* `run_examples()`

check() and test() find no problems.

#### pkgcheck

I am unable to do this locally on my Mac. Here's the error:

```
Error in normalizePath(path, mustWork = TRUE) :
  path[1]="/var/folders/74/jdv3ddd512l0stdy3jzk45b80000gn/T//RtmpjIjxfG/filed972ede8640.txt": No such file or directory
> traceback()
6: normalizePath(path, mustWork = TRUE)
5: brio::read_lines(f_out)
4: pkgstats::ctags_test()
3: withCallingHandlers(expr, message = function(c) if (inherits(c,
       classes)) tryInvokeRestart("muffleMessage"))
2: suppressMessages(pkgstats::ctags_test())
1: pkgcheck()
>
```

This is apparently known because the `normalizePath` documentation says

```
The canonical form of paths may not be what you expect. For example, on macOS absolute paths such as ‘/tmp’ and ‘/var’ are symbolic links
```

#### ropensci-review-bot

Everything here checks ou, including pkgcheck, which ccouldn't be run locally.

Coverage is 93, which is great.

#### Locally-built website

##### /emodnet.wfs/docs/index.html

This is a detailed and well-organized introduction to the {emodnet.wfs} functions.
It is clearly written from the perspective of developers familiar with geospatial or ecological data.
The instructions are easy to follow.

Additional, less technical context would help new users orient themselves to working with this type of data.
An introduction could include things like
* Explanations of the data related terms like "services," "layers," "features", "feature properties", how they relate to each other, and how they relate to more familiar data structures like data frames and columns.
* The data accessed by this package is "geographic vector data." What is important about this format?
* A less technical explanation of WFS that's more approachable than the technical documentaion provided.
* @fmichonneau suggests displaying a map in the READme/this file.
* @fmichonneau suggests recommends adding some context about what kinds of analyses the data sets created with this package are used for.
* What constitutes an observation, and do observations line up across layers or features?

What does CRS stand for? What is the purpose of the transformation?

##### /emodnet.wfs/docs/articles/ecql_filtering.html

Line numbers in corresponding .Rmd file
line 330
starts "There is more that can accomplished"
should be "There is more that can be accomplished"

The paragraph at line 332 recommends using an alternative R package (eurobis) for large data sets.
Since filtering is possible with {emodnet.wfs}, it's not clear from the {eurobis} README why it's better? Is it because of the ability to filter by trait?

##### Using request parameters to limit query results

I was unable to confirm the map from

```
bbox_response %>%
  sf::st_cast(to = "MULTILINESTRING") %>%
  mapview::mapview(burst = TRUE, legend = FALSE)
```

but I did get links for zooming in and out and what sounded like a legend.

line 180
Under the heading for "Non-standard Vendor Parameters", there is a typo in the first line where "additional is missing an i.

### Help pages

It's super helpful that the help pages are crosslinked, so the reader can put the function in it's context in the workflow
(like. the page for `emodnet_get_layer_info` explains that you would have used `emodnet_get_wfs_info`
to get layer names.)

It would be helpful to have more context again about vocabulary like "layers." "features".

The effect of `reduce_layers` would be clearer if I understood layers better and what is being combined.

In `emodnet_get_layers`, the advice about avoiding transferring unnecessary data by subsetting before downloading is great extra strategy that you don't often find in R documentation.

Are these different variables measured at common observation times?

For function `emodnet.wfs`, I'm confused about the term "feature services."  Does "feature" mean "data" in this context, and is that different from the specific features that you can choose to retrieve?

The Value section of the help page for `layer_atribute_inspect` isn't really clear about what the inspection result is. Is it a more detailed documentation then `get_layer_attribute_description`? Or does it calculate summaries?

#### Code

The authors have much more experience with package development than I do, and the code looked great to me.
* Variable and functions names had clear meanings.
* Code was well organized within the `R` directory.
* I didn't find any repetitious code.
* Tasks were broken down into smaller functions that were easier to understand and test.
* The style of the code was readable.

The list of services and URLs is provided as a .csv file with the package.
The authors have probably already thought of this and have a good reason,
like that a similar file is not available from the website.
I would be concerned that the package would need to be updated if new services were added or deleted from the server.

#### Summary

Congratulations to the authors on a great package that will be appreciated by researchers who will be able to easily access this complex data.
It would be helpful to have more documentation describing these complex data structures and the vocabulary used by the package,
both in the vignette/web site and in the help for the individual functions.

Thank you for this opportunity to study good code, learn a couple of new tips, and explore a package in a more broad way than I usually do in my work.
