### 2024-10-30 #653 @MargaretSiple-NOAA
> Is there a way to give mentors in the champions program credit for participating in the review? @fmichonneau made good points that I'd like to add with credit to him.

Hi @LizHareDogs I apologize for not replying to this in the review thread. I could have sworn I wrote you something, but I can't find it! We do not have an official way here to provide credit as a reviewer aside from tagging them in the review thread (which you have done!), but another editor suggested that outside credit like a [LinkedIn recommendation](https://www.linkedin.com/help/linkedin/answer/a541653) is also an appropriate way to record the work they put into mentoring you, and can also be helpful for them when future employers are looking at mentoring experience. I liked [this blog post](https://www.linkedin.com/pulse/how-write-stand-out-linkedin-recommendation-examples-kathryn-rose-) because it has concrete guidelines.

OK, now I am back to reading both of your reviews :)
