### 2024-10-30 #653 @MargaretSiple-NOAA
Hi @maelle - both reviews are in and have insightful comments; thank you both for that. The documentation suggestions from both reviewers will improve the accessibility of the package, especially for new users. I have just a couple small notes on naming conventions/documentation:

- Regarding @robitalec 's suggestion about function names, I will leave it up to you whether you want to change the naming conventions for your functions. The current names are clear and sufficient, and @robitalec 's suggested conventions also make sense.
- Regarding the function descriptions, I agree with both reviewers that the `@description` section can contain more detail and it would help the function documentation to distinguish the titles in the roxygen (for example, three different functions are titled with "Get WFS available layer information").

It is looking good so far and I will pass it off to you to address the reviewers' comments! ⚡
