### 2024-11-04 #653 @maelle
Thank you all for such helpful feedback!!

I'm just back from vacation, and will work on the suggestions ASAP (this week or next). Thank you!!

Regarding acknowledging all reviewers, we can list @fmichonneau along with @robitalec and @LizHareDogs in DESCRIPTION with the "rev" role.
