### 2024-11-05 57df634
docs: remove confusing usage of skimr
