### 2024-11-05 027912c
docs: rephrase to focus less on `View()`
