### 2024-11-05 1916b95
docs: add more about CRS
