+++ vignettes/emodnet.wfs.Rmd
@@ -94,6 +94,8 @@ The column `service_name` shows services available, while  `service_url` has the
 
 ``` r
 seabed_wfs_client <- emodnet_init_wfs_client(service = "seabed_habitats_general_datasets_and_products")
+#> Loading ISO 19139 XML schemas...
+#> Loading ISO 19115 codelists...
 #> ✔ WFS client created successfully
 #> ℹ Service: "https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open/wfs"
 #> ℹ Version: "2.0.0"
@@ -101,18 +103,18 @@ seabed_wfs_client <- emodnet_init_wfs_client(service = "seabed_habitats_general_
 emodnet_get_wfs_info(wfs = seabed_wfs_client)
 #> # A tibble: 72 × 9
 #> # Rowwise: 
-#>    data_source service_name                service_url layer_name title abstract class format
-#>    <chr>       <chr>                       <chr>       <chr>      <chr> <chr>    <chr> <chr> 
-#>  1 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  2 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  3 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  4 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  5 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  6 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  7 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  8 emodnet_wfs seabed_habitats_general_da… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
-#>  9 emodnet_wfs seabed_habitats_general_da… https://ow… carib_eus… 2023… "Output… WFSF… sf    
-#> 10 emodnet_wfs seabed_habitats_general_da… https://ow… biogenic_… Biog… "This l… WFSF… sf    
+#>    data_source service_name    service_url layer_name title abstract class format
+#>    <chr>       <chr>           <chr>       <chr>      <chr> <chr>    <chr> <chr> 
+#>  1 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  2 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  3 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  4 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  5 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  6 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  7 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  8 emodnet_wfs seabed_habitat… https://ow… art17_hab… 2013… "Gridde… WFSF… sf    
+#>  9 emodnet_wfs seabed_habitat… https://ow… carib_eus… 2023… "Output… WFSF… sf    
[…]
