### 2024-11-05 2986a22
docs: remove confusing heading
