+++ vignettes/emodnet.wfs.Rmd
@@ -14,8 +14,7 @@ vignette: >
 
 The package was designed to make EMODnet vector data layers easily accessible in R. The package allows users to query information on and download data from all available [EMODnet Web Feature Service (WFS) endpoints](https://emodnet.ec.europa.eu/geonetwork/emodnet/eng/catalog.search#/home) directly into their R working environment. Data are managed as [`sf` objects](https://r-spatial.github.io/sf/) which are currently the state-of-the-art in handling of vector spatial data in R. The package also allows user to specify the coordinate reference system of imported data.
 
-## Data Product
-### Installation
+## Installation
 
 You can install the development version of emodnet.wfs from GitHub with:
 
@@ -24,7 +23,7 @@ You can install the development version of emodnet.wfs from GitHub with:
 pak::pak("EMODnet/emodnet.wfs")
 ```
 
-### Explore the EMODnet WFS services with R
+## Explore the EMODnet WFS services with R
 
 For this tutorial we will make use of the `sf`,  `dplyr`  and `mapview` packages. The simple features `sf` package is a well known standard for dealing with geospatial vector data. The package `dplyr` is a strong library for data manipulation. This package also loads `magrittr`'s pipe operator `%>%` (you could also use the [base pipe](https://r4ds.hadley.nz/workflow-style.html#sec-pipes)), which allows to write pipelines in R. To visualize geometries,  `mapview` will create quick interactive maps.
 
@@ -94,8 +93,6 @@ The column `service_name` shows services available, while  `service_url` has the
 
 ``` r
 seabed_wfs_client <- emodnet_init_wfs_client(service = "seabed_habitats_general_datasets_and_products")
-#> Loading ISO 19139 XML schemas...
-#> Loading ISO 19115 codelists...
 #> ✔ WFS client created successfully
 #> ℹ Service: "https://ows.emodnet-seabedhabitats.eu/geoserver/emodnet_open/wfs"
 #> ℹ Version: "2.0.0"
