### 2024-11-05 9ff1172
docs: add a bit more about spatial data
