### 2024-11-05 76403c3
docs: more details on geospatial
