### 2024-11-05 51f62fd
docs: better documentation for wfs argument
