### 2024-11-05 439307f
docs: better documentation for wfs argument
