### 2024-11-05 2019463
docs: description for `emodnet_get_layer_info()` and friends
