+++ R/info.R
@@ -78,6 +78,11 @@ emodnet_get_layer_info <- memoise::memoise(.emodnet_get_layer_info)
 }
 #' Get WFS available layer information
 #'
+#' Metadata about data available from the different services:
+#' data (layers) from a data source (service),
+#' metadata on layers from a service,
+#' metadata on layers from all services.
+#'
 #' @param wfs A `WFSClient` R6 object with methods for interfacing an
 #'  OGC Web Feature Service. From [`emodnet_init_wfs_client()`].
 #' @inheritParams emodnet_init_wfs_client
