### 2024-11-05 5354190
docs: description and safer example for `emodnet_wfs()`
