+++ man/emodnet_wfs.Rd
@@ -10,8 +10,10 @@ emodnet_wfs()
 Tibble of available EMODnet Web Feature Services
 }
 \description{
-Available EMODnet Web Feature Services
+Which data sources (services) are available?
 }
 \examples{
+\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 emodnet_wfs()
+\dontshow{\}) # examplesIf}
 }
