### 2024-11-05 9ac3307
docs: description and safer example for `emodnet_wfs()`
