+++ R/emodnet_wfs.R
@@ -6,9 +6,11 @@
 }
 #' Available EMODnet Web Feature Services
 #'
+#' Which data sources (services) are available?
+#'
 #' @return Tibble of available EMODnet Web Feature Services
 #'
-#' @examples
+#' @examplesIf interactive()
 #' emodnet_wfs()
 #' @export
 emodnet_wfs <- memoise::memoise(.emodnet_wfs)
