### 2024-11-05 e6b33a9
docs: description and safer example for `layer_attribute_descriptions()`
