### 2024-11-05 30fa592
docs: description and safer example for `layer_attribute_descriptions()`
