### 2024-11-05 #653 @mpadge
@LizHareDogs Sorry about the error you're seeing with `pkgcheck`. We're aware of some path issues that will be fixed soon. Those are nevertheless [primarily on Windows](https://github.com/ropensci-review-tools/pkgstats/issues/65); issues with paths on Mac should be fixed. If you want to re-run the checks, you could try installing latest version of `pkgcheck`, which should also update `pkgstats` to latest. But only do that if you're interested, or have some other reason - there is no pressing need at all. Thanks!
