### 2024-11-05 b4ccb64
docs: description and safer example for layer_attribute_inspect()
