### 2024-11-05 2da24f8
docs: better manual pages for attributes related functions
