+++ man/roxygen/meta.R
@@ -1,0 +1,5 @@
+list(
+    rd_family_title = list(
+        attributes = "Attributes metadata:"
+    )
+)
