+++ man/layer_attribute_inspect.Rd
@@ -47,3 +47,11 @@ layer_attribute_inspect(
 )
 \dontshow{\}) # examplesIf}
 }
+\seealso{
+Attributes metadata:
+\code{\link{layer_attribute_descriptions}()},
+\code{\link{layer_attributes_get_names}()},
+\code{\link{layer_attributes_summarise}()},
+\code{\link{layer_attributes_tbl}()}
+}
+\concept{attributes}
