### 2024-11-05 815fe8c
docs: better manual pages for attributes related functions
