### 2024-11-07 c1334a0
docs: description and safer example for `emodnet_init_wfs_client()`
