### 2024-11-07 d8522cd
docs: description and safer example for `emodnet_init_wfs_client()`
