### 2024-11-07 5ae594a
docs: safer example for `emodnet_get_wfs_info()`
