+++ man/emodnet_get_wfs_info.Rd
@@ -56,7 +56,7 @@ services from server.
 
 }}
 \examples{
-\dontrun{
+\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 emodnet_get_wfs_info(service = "bathymetry")
 # Query a wfs object
 wfs_bio <- emodnet_init_wfs_client("biology")
@@ -64,5 +64,5 @@ emodnet_get_wfs_info(wfs_bio)
 # Get info for specific layers from wfs object
 layers <- c("mediseh_zostera_m_pnt", "mediseh_posidonia_nodata")
 emodnet_get_layer_info(wfs = wfs_bio, layers = layers)
-}
+\dontshow{\}) # examplesIf}
 }
