### 2024-11-07 27a9ee6
docs: for a better pkgdown ref index
