### 2024-11-07 02f7e66
docs: for a better pkgdown ref index
