### 2024-11-12 341d2cb
docs: list reviewers in DESCRIPTION
