+++ DESCRIPTION
@@ -10,7 +10,13 @@ Authors@R: c(
     person("European Marine Observation Data Network (EMODnet) Biology project", "European Commission's Directorate - General for Maritime Affairs and Fisheries (DG MARE)", , "bio@emodnet.eu", role = "cph"),
     person("VLIZ (VLAAMS INSTITUUT VOOR DE ZEE)", , , "info@vliz.be", role = "fnd"),
     person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = "aut",
-           comment = c(ORCID = "0000-0002-2815-0399"))
+           comment = c(ORCID = "0000-0002-2815-0399")),
+    person("Alec L.", "Robitaille", role = "rev",
+           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>."),
+    person(, role = "rev",
+           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>."),
+    person("François", "Michonneau", , "francois.michonneau@gmail.com", role = "rev",
+           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.")
   )
 Description: Access and interrogate EMODnet (European Marine Observation
     and Data Network) Web Feature Service data through R.
