+++ DESCRIPTION
@@ -12,11 +12,11 @@ Authors@R: c(
     person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = "aut",
            comment = c(ORCID = "0000-0002-2815-0399")),
     person("Alec L.", "Robitaille", role = "rev",
-           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>."),
-    person(, role = "rev",
-           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>."),
+           comment = c(ORCID = "0000-0002-4706-1762", "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.")),
+    person("Liz Hare", role = "rev",
+           comment = c(ORCID = "0000-0002-3978-2543", "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.")),
     person("François", "Michonneau", , "francois.michonneau@gmail.com", role = "rev",
-           comment = "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.")
+           comment = c(ORCID = "0000-0002-9092-966X", "Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>."))
   )
 Description: Access and interrogate EMODnet (European Marine Observation
     and Data Network) Web Feature Service data through R.
