### 2024-11-12 9bd5640
docs: better example escaping, escaped for pkgdown
