+++ man/emodnet_get_wfs_info.Rd
@@ -56,7 +56,7 @@ services from server.
 
 }}
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true") || interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 emodnet_get_wfs_info(service = "bathymetry")
 # Query a wfs object
 wfs_bio <- emodnet_init_wfs_client("biology")
