### 2024-11-12 6126255
docs: better example escaping, escaped for pkgdown
