+++ man/layer_attribute_inspect.Rd
@@ -38,7 +38,7 @@ factor or character variable levels or unique values.
 Inspect layer attributes
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true") || interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 wfs <- emodnet_init_wfs_client(service = "biology")
 layer_attributes_get_names(wfs, layer = "mediseh_zostera_m_pnt")
 layer_attribute_inspect(
