+++ man/emodnet_init_wfs_client.Rd
@@ -25,7 +25,7 @@ OGC Web Feature Service.
 Initialise an EMODnet WFS client
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true") || interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 wfs <- emodnet_init_wfs_client(service = "bathymetry")
 \dontshow{\}) # examplesIf}
 }
