### 2024-11-12 #653 @maelle
@MargaretSiple-NOAA It will take us a few more weeks to finish work on the package as the three of us are juggling other projects, but we have a plan. Should we put this issue on hold?
