### 2024-11-25 861ba71
docs: improve README with more information on spatial data and software
Co-authored-by: Anna Krystalli <annakrystalli@googlemail.com>

Co-authored-by: Salvador Fernández Bejarano <salvador.fernandez@vliz.be>
