+++ README.md
@@ -26,8 +26,11 @@ Services](https://emodnet.ec.europa.eu/en/data-0). [Web Feature services
 change in the way geographic information is created, modified and
 exchanged on the Internet and offer direct fine-grained access to
 geographic information at the feature and feature property level.
-emodnet.wfs aims at offering an user-friendly interface to this rich
-data.
+Features are representation of geographic entities, such as a
+coastlines, marine protected areas, offshore platforms, or fishing
+areas. In WFS, features have geometry (spatial information) and
+attributes (descriptive data). emodnet.wfs aims at offering an
+user-friendly interface to this rich data.
 
 ## Installation and setup
 
@@ -46,35 +49,67 @@ client created successfully”, set the `"emodnet.wfs.quiet"` option to
 options("emodnet.wfs.quiet" = TRUE)
 ```
 
-## Available services
+## Pre-requisites
 
-All available services are contained in the tibble returned by
-`emodnet_wfs()`.
+The emodnet.wfs is designed to be compatible with the modern R
+geospatial stack, in particular output geospatial objects are
+[`sf`](https://r-spatial.github.io/sf/) objects, that is to say, a
+tibble with a geometry list-column.
 
-    #> [1] "data.frame"
-    #> [1] "service_name" "service_url"
-    #>  [1] "bathymetry"                                                     
-    #>  [2] "biology"                                                        
-    #>  [3] "biology_occurrence_data"                                        
-    #>  [4] "chemistry_cdi_data_discovery_and_access_service"                
-    #>  [5] "chemistry_cdi_distribution_observations_per_category_and_region"
-    #>  [6] "chemistry_contaminants"                                         
-    #>  [7] "chemistry_marine_litter"                                        
-    #>  [8] "geology_coastal_behavior"                                       
-    #>  [9] "geology_events_and_probabilities"                               
-    #> [10] "geology_marine_minerals"                                        
-    #> [11] "geology_sea_floor_bedrock"                                      
-    #> [12] "geology_seabed_substrate_maps"                                  
-    #> [13] "geology_submerged_landscapes"                                   
-    #> [14] "human_activities"                                               
-    #> [15] "physics"                                                        
-    #> [16] "seabed_habitats_general_datasets_and_products"                  
-    #> [17] "seabed_habitats_individual_habitat_map_and_model_datasets"
+For users not familiar yet with geospatial data in R, we recommend the
+following resources:
+
+- [Spatial Data Science With Applications in
+  R](https://r-spatial.org/book/) by Edzer Pebesma and Roger Bivand.
+
+- [Geocomputation with R](https://r.geocompx.org/) by Robin Lovelace,
+  Jakub Nowosad and Jannes Muenchow.
+
+In the documentation we assume a basic familiarity with spatial data:
+knowing about coordinates and about projections / [coordinate reference
+systems (CRS)](https://r.geocompx.org/spatial-class#crs-intro).
+
+## Available data sources (services)
+
+All available data sources, called services, are contained in the
+[tibble](https://tibble.tidyverse.org/) returned by `emodnet_wfs()`.
+
+``` r
+library(emodnet.wfs)
+services <- emodnet_wfs()
+class(services)
+#> [1] "data.frame"
+names(services)
+#> [1] "service_name" "service_url"
+services$service_name
+#>  [1] "bathymetry"                                                     
+#>  [2] "biology"                                                        
[…]
