+++ README.Rmd
@@ -26,6 +26,7 @@ options(timeout = 2000)
 
 The goal of emodnet.wfs is to allow interrogation of and access to [EMODnet's, European Marine Observation and Data Network, geographic vector data](https://emodnet.ec.europa.eu/en/emodnet-web-service-documentation#inline-nav-3) in R through the [EMODnet Web Feature Services](https://emodnet.ec.europa.eu/en/data-0). 
 [Web Feature services (WFS)](https://www.ogc.org/publications/standard/wfs/) represent a change in the way geographic information is created, modified and exchanged on the Internet and offer direct fine-grained access to geographic information at the feature and feature property level. 
+Features are representation of geographic entities, such as a coastlines, marine protected areas, offshore platforms, or fishing areas. In WFS, features have geometry (spatial information) and attributes (descriptive data).
 emodnet.wfs aims at offering an user-friendly interface to this rich data.
 
 ## Installation and setup
@@ -44,12 +45,23 @@ set the `"emodnet.wfs.quiet"` option to `TRUE`.
 options("emodnet.wfs.quiet" = TRUE)
 ```
 
-## Available services
+## Pre-requisites
 
-All available services are contained in the tibble returned by `emodnet_wfs()`.
+The emodnet.wfs is designed to be compatible with the modern R geospatial stack, in particular output geospatial objects are [`sf`](https://r-spatial.github.io/sf/) objects, that is to say, a tibble with a geometry list-column.
 
+For users not familiar yet with geospatial data in R, we recommend the following resources:
 
-```{r, echo=FALSE}
+- [Spatial Data Science With Applications in R](https://r-spatial.org/book/) by Edzer Pebesma and Roger Bivand.
+
+- [Geocomputation with R](https://r.geocompx.org/) by Robin Lovelace, Jakub Nowosad and Jannes Muenchow.
+
+In the documentation we assume a basic familiarity with spatial data: knowing about coordinates and about projections / [coordinate reference systems (CRS)](https://r.geocompx.org/spatial-class#crs-intro).
+
+## Available data sources (services)
+
+All available data sources, called services, are contained in the [tibble](https://tibble.tidyverse.org/) returned by `emodnet_wfs()`.
+
+```{r, echo=TRUE}
 library(emodnet.wfs)
 services <- emodnet_wfs()
 class(services)
@@ -59,7 +71,9 @@ services$service_name
 
 To explore available services you can use `View()` or your usual way to explore `data.frames`.
 
-## Create Service Client
+## Initialise a WFS Service Client
+
+A WFS service client is responsible for sending requests to a WFS server and processing the responses to retrieve, display, or analyze geospatial features. As such, initialising a client is the first step to interacting with an EMODnet Web Feature Services.
 
 Specify the service using the `service` argument. 
 
@@ -69,9 +83,11 @@ wfs_bio <- emodnet_init_wfs_client(service = "biology")
 wfs_bio
 ```
 
-## Get WFS Layer info
+## List contents of a WFS: Get layer information from a service client
 
-You can get metadata about the layers available from a service. 
+In the context of a Web Feature Service (WFS), a layer refers to a logical grouping of geographic features that share the same schema (i.e., the same feature type, geometry, and attributes). Layers are the units of data that clients can query, retrieve, and manipulate through a WFS.
+
+You can access information (metadata) about each layer available from an EMODnet WFS with `emodnet_get_wfs_info()`
 
 ```{r}
 emodnet_get_wfs_info(service = "biology")
@@ -99,16 +115,20 @@ emodnet_get_all_wfs_info()
 ```
 
 
-## Get WFS layers
+## Get data from a data source: get layers
 
 You can extract layers directly from a `wfs` object using layer names. 
[…]
