### 2024-11-25 f2a5d62
fix: align output to docs
Co-authored-by: Anna Krystalli <annakrystalli@googlemail.com>

Co-authored-by: Salvador Fernández Bejarano <salvador.fernandez@vliz.be>
