### 2024-11-25 c9fe8b7
docs: add why use eurobis
