+++ vignettes/articles/ecql_filtering.Rmd
@@ -322,6 +322,10 @@ unique(filter_sf3$country)
 
 ### Advanced use
 
-There is more that can be accomplished by using the EMODnet WFS services than downloading data. The emodnet.wfs package is built on top of the [ows4R](https://github.com/eblondel/ows4R) library, meaning that all the functionalities of this package are available for emodnet.wfs. The ows4R returns a special type of R object called R6. You can learn more in Hadley Wickham's chapter on R6 Objects of the [Advance R book](https://adv-r.hadley.nz/r6.html).
+There is more that can be accomplished by using the EMODnet WFS services than downloading data. The `emodnet.wfs` package is built on top of the [ows4R](https://github.com/eblondel/ows4R) library, meaning that all the functionalities of this package are available for emodnet.wfs. The ows4R returns a special type of R object called R6. You can learn more in Hadley Wickham's chapter on R6 Objects of the [Advance R book](https://adv-r.hadley.nz/r6.html).
 
-For instance: it is not efficient to read a large dataset into R just and later subset part of it. This requires longer waiting times and more bandwidth usage, and in very large datasets it would simply not be possible. For instance, all the occurrences data available through the EMODnet Biology portal are stored in [one table](https://www.emodnet-biology.eu/emodnet-data-format): These are approximately 30 millions rows! In this case, we suggest you access the EMODnet Biology occurrence data through the [download toolbox](https://www.emodnet-biology.eu/toolbox/) or the [eurobis R package](https://github.com/lifewatch/eurobis/) instead.
+For instance: it is not efficient to read a large dataset into R just and later subset part of it. This requires longer waiting times and more bandwidth usage, and in very large datasets it would simply not be possible. For instance, all the occurrences data available through the EMODnet Biology portal are stored in [one table](https://www.emodnet-biology.eu/emodnet-data-format): These are approximately 30 millions rows! In this case, we suggest you access the EMODnet Biology occurrence data through the [EMODnet geoviewer](https://emodnet.ec.europa.eu/geoviewer/?layers=6d617269-6e65-696e-666f-000000001510:1:1&basemap=ebwbl&active=12639&bounds=-177.7018404006958,-87.83122047893704,204.7018404006958,100.83122047893704&filters=&projection=EPSG:3857) or the [eurobis R package](https://github.com/lifewatch/eurobis/). This package retrieves occurrences in batches, allowing for large datasets, and uses `viewParams`, a [geoserver-specific filtering feature](https://docs.geoserver.org/main/en/user/data/database/sqlview.html) for faster requests compared to standard OGC or CQL filters. In addition, one of these specific filters allow to query on species traits such as functional groups.
+
+
+
+
