### 2024-11-25 a1d2f73
docs: improve vignette
Co-authored-by: Anna Krystalli <annakrystalli@googlemail.com>

Co-authored-by: Salvador Fernández Bejarano <salvador.fernandez@vliz.be>
