+++ vignettes/emodnet.wfs.Rmd
@@ -1,6 +1,7 @@
 ---
 title: "Accessing and mapping EMODnet data"
 output: rmarkdown::html_vignette
+df_print: "tibble"
 vignette: >
   %\VignetteIndexEntry{API details}
   %\VignetteEngine{knitr::rmarkdown}
@@ -25,65 +26,51 @@ pak::pak("EMODnet/emodnet.wfs")
 
 ## Explore the EMODnet WFS services with R
 
-For this tutorial we will make use of the `sf`,  `dplyr`  and `mapview` packages. The simple features `sf` package is a well known standard for dealing with geospatial vector data. The package `dplyr` is a strong library for data manipulation. This package also loads `magrittr`'s pipe operator `%>%` (you could also use the [base pipe](https://r4ds.hadley.nz/workflow-style.html#sec-pipes)), which allows to write pipelines in R. To visualize geometries,  `mapview` will create quick interactive maps.
+For this tutorial we will make use of the [`sf`](https://r-spatial.github.io/sf/) and [`mapview`](https://r-spatial.github.io/mapview/) packages.
+The simple features `sf` package is a well known standard for dealing with geospatial vector data.
+To visualize geometries, `mapview` will create quick interactive maps.
 
 Run this line to install these packages:
 
 
 ``` r
-install.packages(c("sf", "dplyr", "mapview"))
+install.packages(c("sf", "mapview"))
 ```
 
+EMODnet is organized into [seven thematic lots](https://emodnet.ec.europa.eu/en/emodnet-themes): Bathymetry, Geology, Seabed Habitats, Chemistry, Biology, Physics, and Human Activities, each focusing on a specific aspect of marine data.
+With the emodnet.wfs package, we can explore and combine the data served by the EMODnet thematic lots through [OGC Web Feature Services](https://en.wikipedia.org/wiki/Web_Feature_Service) or WFS.
 
-
-With the emodnet.wfs package, we can explore and combine the data served by the [EMODnet lots ](https://emodnet.ec.europa.eu/en/emodnet-themes) through [OGC Web Feature Services](https://en.wikipedia.org/wiki/Web_Feature_Service) or WFS.
-
-Imagine we are interested in seabed substrates. The first step is to choose what EMODnet lot can provide with these data. For that, we can check the services available with the `emodnet_wfs()` function.
+Imagine we are interested in seabed substrates.
+The first step is to choose what EMODnet thematic lot can provide with these data.
+For that, we can check the services available with the `emodnet_wfs()` function.
 
 
 ``` r
 library(emodnet.wfs)
 library(mapview)
-library(dplyr)
 library(sf)
+#> Error in value[[3L]](cond): Package 'sf' version 1.0.18 cannot be unloaded:
+#>  Error in unloadNamespace(package) : namespace 'sf' is imported by 'ows4R', 'mapview', 'leafem', 'leafpop' so cannot be unloaded
 
-emodnet_wfs()
-#>                                                       service_name
-#> 1                                                       bathymetry
-#> 2                                                          biology
-#> 3                                          biology_occurrence_data
-#> 4                  chemistry_cdi_data_discovery_and_access_service
-#> 5  chemistry_cdi_distribution_observations_per_category_and_region
[…]
