### 2024-11-25 056ed55
docs: explain `reduce_layers` more
