+++ man/emodnet_get_layers.Rd
@@ -48,7 +48,8 @@ Filters with names that do not correspond to any layers are ignored.
 Layers without corresponding filters are returned whole.
 }}
 
-\item{reduce_layers}{whether to reduce output layers to a single \code{sf} object.}
+\item{reduce_layers}{whether to reduce output layers to a single \code{sf} object.
+This only works if the column names are the same.}
 
 \item{...}{additional vendor parameter arguments passed to
 \href{https://docs.geoserver.org/stable/en/user/services/wfs/reference.html#getfeature}{\code{ows4R::GetFeature()}}.
@@ -60,7 +61,8 @@ bigger datasets.}
 If \code{reduce_layers = FALSE} (default), a list of \code{sf}
 objects, one element for each layer. Any layers for which download was
 unsuccessful will be NULL. If \code{reduce_layers = TRUE}, all layers are
-reduced (if possible) to a single \code{sf} containing data for all layers.
+reduced (if possible: if all
+column names are the same) to a single \code{sf} containing data for all layers.
 \code{NULL} layers are ignored. \code{reduce_layers = TRUE} can also be used to return
 an \code{sf} out of a single layer request instead of a list of length 1.
 }
