### 2024-11-25 fe12a42
chore: rename `reduce_layers` to `simplify`
