+++ R/emodnet.wfs-package.R
@@ -1,4 +1,5 @@
 #' @keywords internal
+#' @importFrom lifecycle deprecated
 #' @importFrom rlang .data `%||%`
 "_PACKAGE"
 
