+++ tests/testthat/test-filter.R
@@ -6,7 +6,7 @@ test_that("categorical filters work -- biology", {
       wfs = wfs,
       layers = "mediseh_cymodocea_pnt",
       cql_filter = "country='Israele'",
-      reduce_layers = TRUE
+      simplify = TRUE
     )
   expect_identical(unique(simple_filter_sf$country), "Israele")
 
@@ -14,7 +14,7 @@ test_that("categorical filters work -- biology", {
       wfs = wfs,
       layers = "mediseh_cymodocea_pnt",
       cql_filter = "country='Malta' OR country=='Israele'",
-      reduce_layers = TRUE
+      simplify = TRUE
     )
   })
   expect_identical(unique(or_filter_sf$country), c("Israele", "Malta"))
@@ -28,7 +28,7 @@ test_that("numeric filters work -- biology", {
     num_filter_sf <- emodnet_get_layers(
       wfs = wfs, layers = "mediseh_posidonia_nodata",
       cql_filter = "km>400",
-      reduce_layers = TRUE
+      simplify = TRUE
     )
   })
   expect_gt(min(num_filter_sf$km), 400L)
