+++ tests/testthat/test-vendor-params.R
@@ -4,7 +4,7 @@ test_that("vendor param count works", {
   l_data <- emodnet_get_layers(
     wfs = wfs,
     layers = "mediseh_zostera_m_pnt",
-    count = 1L, reduce_layers = TRUE
+    count = 1L, simplify = TRUE
   )
 
   expect_identical(nrow(l_data), 1L)
