+++ man/emodnet.wfs-package.Rd
@@ -31,6 +31,9 @@ Other contributors:
   \item Thomas J Webb \email{t.j.webb@sheffield.ac.uk} [contributor]
   \item European Marine Observation Data Network (EMODnet) Biology project European Commission's Directorate - General for Maritime Affairs and Fisheries (DG MARE) \email{bio@emodnet.eu} [copyright holder]
   \item VLIZ (VLAAMS INSTITUUT VOOR DE ZEE) \email{info@vliz.be} [funder]
+  \item Alec L. Robitaille (\href{https://orcid.org/0000-0002-4706-1762}{ORCID}) (Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.) [reviewer]
+  \item Liz Hare (\href{https://orcid.org/0000-0002-3978-2543}{ORCID}) (Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.) [reviewer]
+  \item François Michonneau \email{francois.michonneau@gmail.com} (\href{https://orcid.org/0000-0002-9092-966X}{ORCID}) (Reviewed the package (v2.0.2.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/653>.) [reviewer]
 }
 
 }

[…]
