### 2024-11-25 EMODnet/emodnet.wfs#181 @maelle
Collect use cases
Especially data products published by EMODnet
