### 2024-11-25 76bbc7b
docs: add sentence about absence of rate limiting to README
