### 2024-11-25 39ee5a4
feat: add emodnet_thematic_lot column to emodnet_wfs()
