+++ vignettes/emodnet.wfs.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Accessing and mapping EMODnet data"
 output: rmarkdown::html_vignette
-df_print: "tibble"
+df_print: "kable"
 vignette: >
   %\VignetteIndexEntry{API details}
   %\VignetteEngine{knitr::rmarkdown}
@@ -57,25 +57,25 @@ names(services)
 #> [1] "emodnet_thematic_lot" "service_name"         "service_url"
 services[, c("emodnet_thematic_lot", "service_name")]
 #> # A tibble: 17 × 2
-#>    emodnet_thematic_lot     service_name                                        
-#>    <chr>                    <chr>                                               
-#>  1 EMODnet Bathymetry       bathymetry                                          
-#>  2 EMODnet Biology          biology                                             
-#>  3 EMODnet Biology          biology_occurrence_data                             
-#>  4 EMODnet Chemistry        chemistry_cdi_data_discovery_and_access_service     
-#>  5 EMODnet Chemistry        chemistry_cdi_distribution_observations_per_categor…
-#>  6 EMODnet Chemistry        chemistry_contaminants                              
-#>  7 EMODnet Chemistry        chemistry_marine_litter                             
-#>  8 EMODnet Geology          geology_coastal_behavior                            
-#>  9 EMODnet Geology          geology_events_and_probabilities                    
-#> 10 EMODnet Geology          geology_marine_minerals                             
-#> 11 EMODnet Geology          geology_sea_floor_bedrock                           
-#> 12 EMODnet Geology          geology_seabed_substrate_maps                       
-#> 13 EMODnet Geology          geology_submerged_landscapes                        
-#> 14 EMODnet Human Activities human_activities                                    
-#> 15 EMODnet Physics          physics                                             
-#> 16 EMODnet Seabed Habitats  seabed_habitats_general_datasets_and_products       
-#> 17 EMODnet Seabed Habitats  seabed_habitats_individual_habitat_map_and_model_da…
+#>    emodnet_thematic_lot     service_name                                                
+#>    <chr>                    <chr>                                                       
+#>  1 EMODnet Bathymetry       bathymetry                                                  
+#>  2 EMODnet Biology          biology                                                     
+#>  3 EMODnet Biology          biology_occurrence_data                                     
+#>  4 EMODnet Chemistry        chemistry_cdi_data_discovery_and_access_service             
+#>  5 EMODnet Chemistry        chemistry_cdi_distribution_observations_per_category_and_re…
+#>  6 EMODnet Chemistry        chemistry_contaminants                                      
+#>  7 EMODnet Chemistry        chemistry_marine_litter                                     
+#>  8 EMODnet Geology          geology_coastal_behavior                                    
+#>  9 EMODnet Geology          geology_events_and_probabilities                            
[…]
