### 2024-11-26 d3c1ad2
Add new `emodnet_thematic_lot` column to `emodnet_wfs()` and explain what are thematic lots (#182)
* feat: add emodnet_thematic_lot column to emodnet_wfs()

* Update emodnet.wfs.Rmd

* knit

---------

Co-authored-by: Maëlle Salmon <maelle.salmon@yahoo.se>
