+++ R/emodnet_wfs.R
@@ -9,6 +9,17 @@
 #'
 #' @title Which data sources (services) are available?
 #'
+#' @format ## `emodnet_wfs`
+#' \describe{
+#'   \item{emodnet_thematic_lot}{EMODnet disciplinary themes - bathymetry,
+#'   biology, chemistry, geology, human activities, physics and seabed habitats}
+#'   \item{service_name}{Name of the specific service.
+#'   Use in [emodnet_init_wfs_client].}
+#'   \item{service_url}{
+#'   [Web Feature Service (WFS)](https://www.ogc.org/publications/standard/wfs/)
+#'    URL endpoint for accessing the service.}
+#' }
+#'
 #' @return Tibble of available EMODnet Web Feature Services
 #'
 #' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true") || interactive()
