+++ vignettes/emodnet.wfs.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Accessing and mapping EMODnet data"
 output: rmarkdown::html_vignette
-df_print: "tibble"
+df_print: "kable"
 vignette: >
   %\VignetteIndexEntry{API details}
   %\VignetteEngine{knitr::rmarkdown}
@@ -49,33 +49,36 @@ For that, we can check the services available with the `emodnet_wfs()` function.
 library(emodnet.wfs)
 library(mapview)
 library(sf)
-#> Error in value[[3L]](cond): Package 'sf' version 1.0.18 cannot be unloaded:
-#>  Error in unloadNamespace(package) : namespace 'sf' is imported by 'ows4R', 'mapview', 'leafem', 'leafpop' so cannot be unloaded
 
 services <- emodnet_wfs()
-services$service_name
-#>  [1] "bathymetry"                                                     
-#>  [2] "biology"                                                        
-#>  [3] "biology_occurrence_data"                                        
-#>  [4] "chemistry_cdi_data_discovery_and_access_service"                
-#>  [5] "chemistry_cdi_distribution_observations_per_category_and_region"
-#>  [6] "chemistry_contaminants"                                         
-#>  [7] "chemistry_marine_litter"                                        
-#>  [8] "geology_coastal_behavior"                                       
-#>  [9] "geology_events_and_probabilities"                               
-#> [10] "geology_marine_minerals"                                        
-#> [11] "geology_sea_floor_bedrock"                                      
-#> [12] "geology_seabed_substrate_maps"                                  
-#> [13] "geology_submerged_landscapes"                                   
-#> [14] "human_activities"                                               
-#> [15] "physics"                                                        
-#> [16] "seabed_habitats_general_datasets_and_products"                  
-#> [17] "seabed_habitats_individual_habitat_map_and_model_datasets"
+class(services)
+#> [1] "tbl_df"     "tbl"        "data.frame"
+names(services)
+#> [1] "emodnet_thematic_lot" "service_name"         "service_url"
+services[, c("emodnet_thematic_lot", "service_name")]
+#> # A tibble: 17 × 2
+#>    emodnet_thematic_lot     service_name                                                
+#>    <chr>                    <chr>                                                       
+#>  1 EMODnet Bathymetry       bathymetry                                                  
+#>  2 EMODnet Biology          biology                                                     
+#>  3 EMODnet Biology          biology_occurrence_data                                     
+#>  4 EMODnet Chemistry        chemistry_cdi_data_discovery_and_access_service             
+#>  5 EMODnet Chemistry        chemistry_cdi_distribution_observations_per_category_and_re…
+#>  6 EMODnet Chemistry        chemistry_contaminants                                      
+#>  7 EMODnet Chemistry        chemistry_marine_litter                                     
+#>  8 EMODnet Geology          geology_coastal_behavior                                    
[…]
