### 2024-12-06 #653 @maelle
Thanks a lot to the reviewers @robitalec, @lizHareDogs, @fmichonneau and to the editor @MargaretSiple-NOAA for extremely useful feedback. We [listed all reviewers](https://github.com/EMODnet/emodnet.wfs/pull/179) in the `DESCRIPTION` file.

## Response to @robitalec's review

### Function naming

> My related suggestion is to consider including a dictionary of this scheme somewhere in the README or introductory vignette. I am not sure exactly where/if this already exists in the broader EMODnet documentation but I think a description of each family of functions wfs/layer/attributes would be useful for new users getting started with the package.

We’ve decided to keep the current naming that we had chosen to align with the OGC standards. We’ve made many improvements to the documentation (README including heading titles, vignettes, manual pages) to ensure (hopefully!) the package becomes more user-friendly.

### Function documentation

> The documentation for this package is not extensively detailed. For example, the description sections for R functions are missing except for 1 function (emodnet_get_layers). I would suggest including a paragraph to describe each exported function so the user sees more than the title and list of parameters when they first open the help pages. I would also suggest including the name of the function in the parameters when an argument for a function is the output from another function, eg. in [R/info.R](https://github.com/EMODnet/emodnet.wfs/blob/d98df3354e7abb1528c3eb6cc5bdd64e6bc8b90b/R/info.R#L81-L82): @param wfs …, from emodnet_init_wfs_client.

We have improved all manual pages, in particular we have made sure the titles of functions are more user-facing (with words like data sources, datasets, variables instead of web services, layers, attributes) with the description using the more specific words. We also improved the titles of sections on the pkgdown reference page. https://github.com/EMODnet/emodnet.wfs/pull/177

### Introductory vignette

> I appreciated the suggestion from @MargaretSiple-NOAA but I think one step further for renaming the introductory vignette is to highlight that it is the introductory vignette with something like “getting started” or “introduction to” etc. At the moment, the introductory vignette on the pkgdown site is nestled between the other two vignettes making it unclear which vignette a user should start with.

We changed the title and position of the vignette in the [navbar](https://github.com/EMODnet/emodnet.wfs/commit/38ac5bdf60feb4e7eb7bc0a413fd3a96846d0c6f)

> The word “lot” is used a couple of times but it is unclear what this word refers to with respect to emodnet.wfs functions (wfs/layers/attributes).

We have rephrase the uses of “EMODnet lot” to “EMODnet thematic lot”. We have explained how EMODnet is organized in 7 thematic lots in the README.md and get started vignette. We have also added a new column to the emodnet_wfs() tibble `emodnet_thematic_lot` to specify the emodnet lot that manages the service.

> Consider using {knitr} tables or similar functions to tidy up the printing of outputs.

We show data.frames as tibbles in https://emodnet.github.io/emodnet.wfs/articles/emodnet.wfs.html

> [This chunk](https://github.com/EMODnet/emodnet.wfs/blob/d98df3354e7abb1528c3eb6cc5bdd64e6bc8b90b/vignettes/emodnet.wfs.Rmd#L173-L186) feels like a pretty non-introductory chunk of code. Not sure if it's required or could be simplified.

We have simplified the code by requesting instead the data as geojson, and explained the reasoning for it. This is a known feature / issue of OGC services and we linked to sources explaining the feature.

> There is an error related to mapview/webshot in the pkgdown/rendered vignette:

We [fixed this bug](https://github.com/EMODnet/emodnet.wfs/commit/c2eef34ba04dae939fd954c18f751bc84aff9a57).

### Returned objects

> The argument "reduce_layers" could use more explanation. The documentation suggests reducing layers will be attempted, but what is an example of reasons which reducing might fail?

> Why does a single layer return a list? I think it's intuitive to match the names of multiple layers requested to a named list but wouldn't a user specifying a single layer expect an sf object to be directly returned?

We have renamed the argument to “simplify” to align it with base R functions, and we have clarified in the manual that it only works if column names are the same.

### Minor comments

> Given the overlap in authorship, if the authors are aware of how EMODnetWCS differs from emodnet.wfs or what the roadmap looks like for EMODnetWCS, it might be useful to include it in the README under other web services. Are [these](https://github.com/EMODnet/EMODnet-Biology-Guidance?tab=readme-ov-file#r-client-for-emodnet-biology-webservices) also additional tools to reference there? Could you add more precision to how "these three ways to access EMODnet complement each other"?

We have made the section about other packages more extensive: https://emodnet.github.io/emodnet.wfs/#unlock-the-full-potential-of-the-emodnet-web-services-access-raster-and-gridded-datasets

> Codecov badge present in README twice, pointing to the root and tree/main, giving different code coverage percentages.

We [removed the duplication](https://github.com/EMODnet/emodnet.wfs/commit/3be8e11c1e2bb4760cdf1b20136f2891ac8593bd).

> The authors could use `cffr::cff_write()` to generate a citation file.

[Now done](https://github.com/EMODnet/emodnet.wfs/commit/2c2e9df8a53ba1a0ed5d2969d4aed3a3e1551bff).

> Are there any rate limits/rules for the API? If so, could these be mentioned in the README and vignettes?

We have added the sentence “The use of the EMODnet Web Feature Services is not subjet to rate limiting at the moment.” to the README.

> The skimr package is listed in the Suggests only to be used once in the ECQL filtering vignette. Is it a necessary dependency? I find reading through the vignette that the single skimr line returns an unexpected/different/confusing output to the previous outputs.

We ended up removing the line using skimr.

> Some broken links detected with {urlchecker}.

We fixed the links in a few commits. Urlchecker now finds no broken link.

## Response to @LizHareDogs's review

### README.md

#### Readthrough

> From the perspective of someone who has not worked with this kind of data,  it would be helpful if the introduction (which starts at line 21) could provide a little more context. What is geographic
vector data? I'm new to this kind of data so it would be helpful to know in a less technical way what it is.

We've added context and pre-requisites to the README.

> Line 26
> "though" should be "through"

We fixed the typo.

> line 78
> Not sure if I can use "View"

We replace the code with the sentence “To explore available services you can use `View()` or your usual way to explore `data.frames`.”

> line 184
> What does CRS stand for (it's mentioned a few times later)?

We added the definition in two vignettes.

> Line 313
> what is an sf object? What is the difference between this and how the data were retrieved before?

We added the context to the README in https://emodnet.github.io/emodnet.wfs/index.html#pre-requisites

#### Following instructions on Mac OS using Emacs, Emacs SPeaks Statistics

> Line 78
> I am not able to evaluate the View command on my system. This is not a problem with the package.
> Otherwise, sample commands work as expected.

We have removed the exclusive mention of `View() as a way to explore data.frames.

### Vignette

> All the commands in the vignette work for me.

> Line 24 has the heading "data product." The meaning of this isn't clear to me. Will this be instructions on how to access a data product or how to create a data product?

We removed the confusing heading. “Data product” is an internal phrase for EMODnet team members, irrelevant for users of the package.

> Line 109 has this command:
> map <- mapview(habitats_directive_layers, zcol = "habitat_description", burst = TRUE)

> I am able to run this command, which opens a browser. I can't check the map with the screen reader, but the text on the > screen seems appropriate.

> In the rendered vignette at
> the following error is displayed for the mapview command:

We have fixed the bug that lead to this chunk not working (we had not registered the
dependency on webshot).

### Locally-built website

#### /emodnet.wfs/docs/index.html

> This is a detailed and well-organized introduction to the {emodnet.wfs} functions.
> It is clearly written from the perspective of developers familiar with geospatial or ecological data.
> The instructions are easy to follow.

> Additional, less technical context would help new users orient themselves to working with this type of data.
> An introduction could include things like
> * Explanations of the data related terms like "services," "layers," "features", "feature properties", how they relate to each > other, and how they relate to more familiar data structures like data frames and columns.
> * The data accessed by this package is "geographic vector data." What is important about this format?
> * A less technical explanation of WFS that's more approachable than the technical documentaion provided.
> * @fmichonneau suggests displaying a map in the READme/this file.
> * @fmichonneau suggests recommends adding some context about what kinds of analyses the data sets created with
> this package are used for.
> * What constitutes an observation, and do observations line up across layers or features?

> What does CRS stand for? What is the purpose of the transformation?

We have improved the manual pages, README, vignettes to ensure all terms like “layers” are accompanied by more beginner-friendly explanations. We have listed pre-requisites and resources in the README.
We prefer only showing a map in the introductory vignette.
We do not have published use cases yet but opened an issue to remind us to list them once available: https://github.com/EMODnet/emodnet.wfs/issues/181

#### /emodnet.wfs/docs/articles/ecql_filtering.html

> Line numbers in corresponding .Rmd file
> line 330
> starts "There is more that can accomplished"
> should be "There is more that can be accomplished"

We fixed the typo.

> The paragraph at line 332 recommends using an alternative R package (eurobis) for large data sets.
> Since filtering is possible with {emodnet.wfs}, it's not clear from the {eurobis} README why it's better? Is it because of > the ability to filter by trait?

We have improved the filtering vignette with an explanation of why to use eurobis.

#### Using request parameters to limit query results

> I was unable to confirm the map from
> (...)
> but I did get links for zooming in and out and what sounded like a legend.

> line 180
> Under the heading for "Non-standard Vendor Parameters", there is a typo in the first line where "additional is missing an i.

We fixed the typo.

### Help pages

> It's super helpful that the help pages are crosslinked, so the reader can put the function in it's context in the workflow
(like. the page for `emodnet_get_layer_info` explains that you would have used `emodnet_get_wfs_info`
to get layer names.)

> It would be helpful to have more context again about vocabulary like "layers." "features".

We have improved help pages. In particular the reference index now features “easier” terms near the official terms: https://emodnet.github.io/emodnet.wfs/reference/index.html

> The effect of `reduce_layers` would be clearer if I understood layers better and what is being combined.

We have renamed the parameter to `simplify` and improved its documentation.

> In `emodnet_get_layers`, the advice about avoiding transferring unnecessary data by subsetting before downloading is great extra strategy that you don't often find in R documentation.

> Are these different variables measured at common observation times?

In short: no. The only thing in common between the data available through this package is that they are in vector format, hence they can be points, lines, polygons etc. Depending on the dataset (layer) you are looking at, there might be a variable (attribute) that specifies time, but not all datasets (layer) will have a time dimension. In any case, when a time dimension is specified, EMODnet follows the ISO 8601 standard. Except if you are looking at multidimensional gridded data in a CF / COARDS compliant netcdf file, typically in ERDDAP, which the Unix Time format, normally as “days since January 1, 1970".

> For function `emodnet.wfs`, I'm confused about the term "feature services."  Does "feature" mean "data" in this context, and is that different from the specific features that you can choose to retrieve?

We have improved context in the README and headers of manual pages. Example: https://emodnet.github.io/emodnet.wfs/reference/emodnet_get_wfs_info.html

> The Value section of the help page for `layer_atribute_inspect` isn't really clear about what the inspection result is. Is it a more detailed documentation then `get_layer_attribute_description`? Or does it calculate summaries?

We have added cross-references between manual pages of functions that help examine attributes, and we have made sure there are evaluated examples in the pkgdown manual pages.

### Code

> The authors have much more experience with package development than I do, and the code looked great to me.
> * Variable and functions names had clear meanings.
> * Code was well organized within the `R` directory.
> * I didn't find any repetitious code.
> * Tasks were broken down into smaller functions that were easier to understand and test.
> * The style of the code was readable.

> The list of services and URLs is provided as a .csv file with the package.
> The authors have probably already thought of this and have a good reason,
> like that a similar file is not available from the website.
> I would be concerned that the package would need to be updated if new services were added or deleted from the server.

We agree that it is a limitation, but before making a new service available to users of the package we would like to test it.

## Conclusion (for now :smile_cat: )

Thanks again to the reviewers and editor, we feel our package already became much stronger as a result of the feedback we received.
