### 2025-01-15 #653 @robitalec
## Reviewer Response

(Apologies for my delay).

Thank you @maelle for your detailed response to our reviews.

Ok for keeping the current function naming, good to know about the choice relating to OGC standards.

I see a number of excellent improvements to:

- README
- manual pages
- vignette section about other packages
- clarification and explanation of various emodnet terms eg. "lot", "service clients", "layer info"
- spatial R pre-requisites

Thank you @maella for carefully considering my comments on this excellent package and @MargaretSiple-NOAA for editing this package review and helping along the way.

#### Final approval (post-review)

Estimated hours spent reviewing: 5 (4.5 + 0.5)
