### 2025-01-20 #653 @LizHareDogs
I approve!

Thanks very much to the authors for all the editing of the documentation! I think it will be a lot easier for new users to get started and I appreciate the time and work. Congratulations on a great package.
