### 2025-01-23 #653 @MargaretSiple-NOAA
It looks like the issue with periods in package names has been [resolved](https://github.com/ropensci-org/buffy/pull/110)... going to try again.
