### 2025-01-27 #653 @maelle
I worked through the checklist, although we haven't transferred emodnet.wfs and won't (it's now listed in the registry, I added it to https://github.com/ropensci/roregistry/blob/gh-pages/info/not_transferred.json thanks to my insider knowledge :wink:)

We're planning a release soon but need a bug to be fixed upstream in ows4R so that we can remove a hot fix we made in emodnet.wfs.
