### 2025-01-27 #653 @maelle
(I'm working on improving the docs for those without insider knowledge 😇 https://github.com/ropensci/software-review/pull/682)
